Update mount design
The first version's spring is too weak and the device is very loose on
the DIN rail. The spring thickness was increased and the slider position
is modified so the spring will be under tension when it's on the DIN
rail. Hopefully this tension will make the device more stable.